refactored command packages
