+++ src/main/java/hy/tmc/cli/frontend_communication/Commands/Authenticate.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package hy.tmc.cli.frontend_communication.Commands.CommandLineClientCommands;
+package hy.tmc.cli.frontend_communication.Commands;
 
 import hy.tmc.cli.Configuration.ClientData;
 import static hy.tmc.cli.backendCommunication.URLCommunicator.*;
