+++ src/main/java/hy/tmc/cli/frontend_communication/Commands/ListCourses.java
@@ -3,41 +3,33 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-
-package hy.tmc.cli.frontend_communication.Commands.CommandLineClientCommands;
+package hy.tmc.cli.frontend_communication.Commands;
 
 import hy.tmc.cli.frontend_communication.Commands.Command;
 import hy.tmc.cli.frontend_communication.FrontendListener;
-import hy.tmc.cli.frontend_communication.Server.Server;
+import hy.tmc.cli.frontend_communication.Server.ProtocolException;
 import hy.tmc.cli.logic.Logic;
-import java.io.IOException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 /**
  *
- * @author ilari
+ * @author chang
  */
-public class ReplyToPing extends Command {
-    
-    private final String answer = "pong";
+public class ListCourses extends Command {
 
-    public ReplyToPing(FrontendListener front, Logic backend) {
+    public ListCourses(FrontendListener front, Logic backend) {
         super(front, backend);
     }
 
     @Override
     protected void functionality() {
-        this.frontend.printLine(answer);
     }
 
     @Override
     public void setParameter(String key, String value) {
     }
-    
+
     @Override
-    public void checkData(){
-        
+    public void checkData() throws ProtocolException {
     }
     
 }
